+++ packages/vinxi/lib/build.js
@@ -56,7 +56,7 @@ export async function createBuild(app, buildConfig, configFile) {
 			}
 
 			await withLogger({ router, requestId: "build" }, async () => {
-				await createRouterBuild(app, router);
+				await createRouterBuild(app, router, buildConfig.mode);
 			});
 		}
 
@@ -91,7 +91,7 @@ export async function createBuild(app, buildConfig, configFile) {
 	for (const router of app.config.routers) {
 		if (router.type !== "static" && router.build !== false) {
 			await withLogger({ router, requestId: "build" }, async () => {
-				await createRouterBuildInWorker(app, router, configFile);
+				await createRouterBuildInWorker(app, router, buildConfig.mode, configFile);
 			});
 		}
 	}
@@ -397,19 +397,27 @@ async function createViteBuild(config) {
 	return output;
 }
 
-async function createRouterBuildInWorker(app, router, configFile) {
+/**
+ *
+ * @param {import("./app.js").App} app
+ * @param {import("./router-mode.js").Router} router
+ * @param {string} [mode]
+ * @param {string} [configFile]
+ */
+async function createRouterBuildInWorker(app, router, mode, configFile) {
 	const sh = await import("../runtime/sh.js");
 	const { fileURLToPath } = await import("url");
 	await sh.default`node ${fileURLToPath(
 		new URL("../bin/cli.mjs", import.meta.url).href,
-	)} build --router=${router.name} ${configFile ? `--config=${configFile}` : ""}`;
+	)} build --router=${router.name} ${mode ? `--mode=${mode}` : ""} ${configFile ? `--config=${configFile}` : ""}`;
 }
 /**
  *
  * @param {import("./app.js").App} app
  * @param {import("./router-mode.js").Router} router
+ * @param {string} [mode]
  */
-async function createRouterBuild(app, router) {
+async function createRouterBuild(app, router, mode) {
 	console.log("\n");
 	console.log(c.green(`📦 Compiling ${router.name} router...`));
 	await app.hooks.callHook("app:build:router:start", { app, router });
@@ -422,7 +430,7 @@ async function createRouterBuild(app, router) {
 		await createViteBuild({
 			app: app,
 			root: router.root,
-			mode: app.config.mode,
+			mode,
 			build: {
 				ssr: true,
 				ssrManifest: true,
@@ -503,6 +511,7 @@ async function createRouterBuild(app, router) {
 				},
 			},
 		],
+		mode,
 	};
 
 	await app.hooks.callHook("app:build:router:vite:config", {