+++ test/templates/multi-spa/CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### Patch Changes
 
+- Updated dependencies [b85ab82c]
+  - vinxi@0.5.2
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [db9b7747]
   - vinxi@0.5.1